fix: Handle EXIF orientation in uploaded gallery images (#1617, #1694).
Photoswipe needs correct width and height information for all images,
or it will show them stretched to an incorrect aspect ratio. When an
image is uploaded with an EXIF orientation of 6 or 8 (rotated +90 or
-90 degrees), the width and height need to be swapped.

Add helpers/rotation-aware-sizeof.js, which wraps the image-size
module and swaps width and height if necessary.